+++ modules/canbus/vehicle/gem/protocol/turn_cmd_63.cc
@@ -60,7 +60,7 @@ void Turncmd63::set_p_turn_signal_cmd(
   uint8_t x = turn_signal_cmd;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 0, 8);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
 }
 
 }  // namespace gem
