+++ modules/canbus/vehicle/lexus/protocol/cruise_control_buttons_cmd_108.cc
@@ -77,7 +77,7 @@ void Cruisecontrolbuttonscmd108::set_p_cruise_control_button(
   uint8_t x = cruise_control_button;
 
   Byte to_set(data + 1);
-  to_set.set_value(x, 0, 8);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
 }
 
 Cruisecontrolbuttonscmd108* Cruisecontrolbuttonscmd108::set_ignore_overrides(
@@ -94,7 +94,7 @@ void Cruisecontrolbuttonscmd108::set_p_ignore_overrides(uint8_t* data,
   uint8_t x = ignore_overrides;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 1, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 1, 1);
 }
 
 Cruisecontrolbuttonscmd108* Cruisecontrolbuttonscmd108::set_clear_override(
@@ -111,7 +111,7 @@ void Cruisecontrolbuttonscmd108::set_p_clear_override(uint8_t* data,
   uint8_t x = clear_override;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 2, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 2, 1);
 }
 
 Cruisecontrolbuttonscmd108* Cruisecontrolbuttonscmd108::set_enable(
@@ -127,7 +127,7 @@ void Cruisecontrolbuttonscmd108::set_p_enable(uint8_t* data, bool enable) {
   uint8_t x = enable;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 0, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 1);
 }
 
 Cruisecontrolbuttonscmd108* Cruisecontrolbuttonscmd108::set_clear_faults(
@@ -144,7 +144,7 @@ void Cruisecontrolbuttonscmd108::set_p_clear_faults(uint8_t* data,
   uint8_t x = clear_faults;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 3, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 3, 1);
 }
 
 }  // namespace lexus
